refactor(forks/lstar): collapse containers package into a single module
Replace the lstar containers/ package and its attestation, block, and
state subpackages with one containers.py file holding every container.
Update all import sites to pull from lean_spec.forks.lstar.containers.
