+++ packages/testing/src/consensus_testing/genesis.py
@@ -1,8 +1,13 @@
 """Consensus layer pre-state generation."""
 
-from lean_spec.forks.lstar.containers.block import AggregatedAttestations, Block, BlockBody
-from lean_spec.forks.lstar.containers.state import State, Validators
-from lean_spec.forks.lstar.containers.validator import Validator
+from lean_spec.forks.lstar.containers import (
+    AggregatedAttestations,
+    Block,
+    BlockBody,
+    State,
+    Validator,
+    Validators,
+)
 from lean_spec.forks.lstar.spec import LstarSpec
 from lean_spec.subspecs.ssz.hash import hash_tree_root
 from lean_spec.types import Bytes52, Slot, Uint64, ValidatorIndex
