+++ packages/testing/src/consensus_testing/test_fixtures/fork_choice.py
@@ -11,14 +11,13 @@ from typing import ClassVar, Self
 
 from pydantic import Field, model_validator
 
-from lean_spec.forks.lstar.containers.block import (
+from lean_spec.forks.lstar.containers import (
+    AggregatedAttestations,
     Block,
     BlockBody,
+    State,
+    Validators,
 )
-from lean_spec.forks.lstar.containers.block.types import (
-    AggregatedAttestations,
-)
-from lean_spec.forks.lstar.containers.state import State, Validators
 from lean_spec.forks.lstar.spec import LstarSpec
 from lean_spec.subspecs.chain.clock import Interval
 from lean_spec.subspecs.ssz import hash_tree_root
