+++ packages/testing/src/consensus_testing/test_types/gossip_aggregated_attestation_spec.py
@@ -2,10 +2,12 @@
 
 from __future__ import annotations
 
-from lean_spec.forks.lstar.containers.attestation import AttestationData
-from lean_spec.forks.lstar.containers.attestation.attestation import SignedAggregatedAttestation
-from lean_spec.forks.lstar.containers.block.block import Block
-from lean_spec.forks.lstar.containers.state import State
+from lean_spec.forks.lstar.containers import (
+    AttestationData,
+    Block,
+    SignedAggregatedAttestation,
+    State,
+)
 from lean_spec.subspecs.xmss.aggregation import TypeOneMultiSignature
 from lean_spec.types import (
     ByteList512KiB,
