+++ packages/testing/src/consensus_testing/test_types/state_expectation.py
@@ -3,13 +3,13 @@
 from collections.abc import Callable
 from typing import Any, ClassVar
 
-from lean_spec.forks.lstar.containers.block.block import Block
-from lean_spec.forks.lstar.containers.state import State
-from lean_spec.forks.lstar.containers.state.types import (
+from lean_spec.forks.lstar.containers import (
+    Block,
     HistoricalBlockHashes,
     JustificationRoots,
     JustificationValidators,
     JustifiedSlots,
+    State,
 )
 from lean_spec.types import Bytes32, CamelModel, Slot
