+++ packages/testing/src/consensus_testing/test_types/step_types.py
@@ -4,11 +4,11 @@ from typing import Annotated, Any, Literal, Union
 
 from pydantic import ConfigDict, Field, PrivateAttr, field_serializer, model_validator
 
-from lean_spec.forks.lstar.containers.attestation import (
+from lean_spec.forks.lstar.containers import (
+    Block,
     SignedAggregatedAttestation,
     SignedAttestation,
 )
-from lean_spec.forks.lstar.containers.block.block import Block
 from lean_spec.types import CamelModel
 
 from .block_spec import BlockSpec
