+++ packages/testing/src/consensus_testing/test_types/store_checks.py
@@ -2,8 +2,7 @@
 
 from typing import Literal
 
-from lean_spec.forks.lstar.containers import AttestationData
-from lean_spec.forks.lstar.containers.block.block import Block
+from lean_spec.forks.lstar.containers import AttestationData, Block
 from lean_spec.forks.lstar.spec import LstarSpec
 from lean_spec.forks.lstar.store import Store
 from lean_spec.subspecs.chain.clock import Interval
