+++ packages/testing/src/consensus_testing/test_types/utils.py
@@ -2,7 +2,7 @@
 
 from __future__ import annotations
 
-from lean_spec.forks.lstar.containers.block.block import Block
+from lean_spec.forks.lstar.containers import Block
 from lean_spec.subspecs.ssz.hash import hash_tree_root
 from lean_spec.types import Bytes32, Checkpoint, Slot
