--- src/lean_spec/forks/lstar/containers/attestation/__init__.py
@@ -1,17 +1,0 @@
-"""Attestation containers and related types for the Lean spec."""
-
-from .attestation import (
-    AggregatedAttestation,
-    Attestation,
-    AttestationData,
-    SignedAggregatedAttestation,
-    SignedAttestation,
-)
-
-__all__ = [
-    "AggregatedAttestation",
-    "Attestation",
-    "AttestationData",
-    "SignedAggregatedAttestation",
-    "SignedAttestation",
-]
