--- src/lean_spec/forks/lstar/containers/block/__init__.py
@@ -1,17 +1,0 @@
-"""Block containers and related types for the Lean Ethereum consensus specification."""
-
-from .block import (
-    Block,
-    BlockBody,
-    BlockHeader,
-    SignedBlock,
-)
-from .types import AggregatedAttestations
-
-__all__ = [
-    "Block",
-    "BlockBody",
-    "BlockHeader",
-    "SignedBlock",
-    "AggregatedAttestations",
-]
