--- src/lean_spec/forks/lstar/containers/state/__init__.py
@@ -1,19 +1,0 @@
-"""State container and related types for the Lean Ethereum consensus specification."""
-
-from ..validator import Validators
-from .state import State
-from .types import (
-    HistoricalBlockHashes,
-    JustificationRoots,
-    JustificationValidators,
-    JustifiedSlots,
-)
-
-__all__ = [
-    "HistoricalBlockHashes",
-    "JustificationRoots",
-    "JustificationValidators",
-    "JustifiedSlots",
-    "State",
-    "Validators",
-]
