--- src/lean_spec/forks/lstar/containers/state/state.py
@@ -1,51 +1,0 @@
-"""State Container for the Lean Ethereum consensus specification."""
-
-from lean_spec.forks.lstar.containers.block import BlockHeader
-from lean_spec.forks.lstar.containers.config import Config
-from lean_spec.forks.lstar.containers.state.types import (
-    HistoricalBlockHashes,
-    JustificationRoots,
-    JustificationValidators,
-    JustifiedSlots,
-)
-from lean_spec.forks.lstar.containers.validator import Validators
-from lean_spec.types import Checkpoint, Container, Slot
-
-
-class State(Container):
-    """The main consensus state object."""
-
-    # Configuration
-    config: Config
-    """The chain's configuration parameters."""
-
-    # Slot and block tracking
-    slot: Slot
-    """The current slot number."""
-
-    latest_block_header: BlockHeader
-    """The header of the most recent block."""
-
-    # Checkpoints
-    latest_justified: Checkpoint
-    """The latest justified checkpoint."""
-
-    latest_finalized: Checkpoint
-    """The latest finalized checkpoint."""
-
-    # Historical data
-    historical_block_hashes: HistoricalBlockHashes
-    """A list of historical block root hashes."""
-
-    justified_slots: JustifiedSlots
-    """A bitfield indicating which historical slots were justified."""
-
-    validators: Validators
-    """Registry of validators tracked by the state."""
-
-    # Justification tracking (flattened for SSZ compatibility)
-    justifications_roots: JustificationRoots
-    """Roots of justified blocks."""
-
-    justifications_validators: JustificationValidators
-    """A bitlist of validators who participated in justifications."""
