--- src/lean_spec/forks/lstar/containers/validator.py
@@ -1,32 +1,0 @@
-"""Validator container for the Lean Ethereum consensus specification."""
-
-from lean_spec.subspecs.chain.config import VALIDATOR_REGISTRY_LIMIT
-from lean_spec.subspecs.xmss.containers import PublicKey
-from lean_spec.types import Bytes52, Container, SSZList, ValidatorIndex
-
-
-class Validator(Container):
-    """Represents a validator's static metadata and operational interface."""
-
-    attestation_pubkey: Bytes52
-    """XMSS public key for signing attestations."""
-
-    proposal_pubkey: Bytes52
-    """XMSS public key for signing proposer attestations in blocks."""
-
-    index: ValidatorIndex = ValidatorIndex(0)
-    """Validator index in the registry."""
-
-    def get_attestation_pubkey(self) -> PublicKey:
-        """Get the XMSS public key used for attestation verification."""
-        return PublicKey.decode_bytes(bytes(self.attestation_pubkey))
-
-    def get_proposal_pubkey(self) -> PublicKey:
-        """Get the XMSS public key used for proposer attestation verification."""
-        return PublicKey.decode_bytes(bytes(self.proposal_pubkey))
-
-
-class Validators(SSZList[Validator]):
-    """Validator registry tracked in the state."""
-
-    LIMIT = int(VALIDATOR_REGISTRY_LIMIT)
