+++ tests/consensus/lstar/state_transition/test_finalization.py
@@ -9,7 +9,7 @@ from consensus_testing import (
     generate_pre_state,
 )
 
-from lean_spec.forks.lstar.containers.state.types import (
+from lean_spec.forks.lstar.containers import (
     JustificationRoots,
     JustificationValidators,
     JustifiedSlots,
