+++ tests/consensus/lstar/state_transition/test_genesis.py
@@ -16,16 +16,17 @@ from consensus_testing import (
     generate_pre_state,
 )
 
-from lean_spec.forks.lstar.containers.block import BlockBody
-from lean_spec.forks.lstar.containers.block.types import AggregatedAttestations
-from lean_spec.forks.lstar.containers.state import State
-from lean_spec.forks.lstar.containers.state.types import (
+from lean_spec.forks.lstar.containers import (
+    AggregatedAttestations,
+    BlockBody,
     HistoricalBlockHashes,
     JustificationRoots,
     JustificationValidators,
     JustifiedSlots,
+    State,
+    Validator,
+    Validators,
 )
-from lean_spec.forks.lstar.containers.validator import Validator, Validators
 from lean_spec.forks.lstar.spec import LstarSpec
 from lean_spec.subspecs.chain.config import VALIDATOR_REGISTRY_LIMIT
 from lean_spec.subspecs.ssz.hash import hash_tree_root
