+++ tests/interop/helpers/node_runner.py
@@ -13,8 +13,7 @@ from dataclasses import dataclass, field
 from typing import cast
 
 from lean_spec.forks.lstar import Store
-from lean_spec.forks.lstar.containers import Validator
-from lean_spec.forks.lstar.containers.state import Validators
+from lean_spec.forks.lstar.containers import Validator, Validators
 from lean_spec.forks.lstar.spec import LstarSpec
 from lean_spec.subspecs.chain.config import ATTESTATION_COMMITTEE_COUNT
 from lean_spec.subspecs.networking import PeerId
