+++ tests/lean_spec/forks/lstar/forkchoice/test_attestation_target.py
@@ -9,9 +9,9 @@ from lean_spec.forks.lstar import Store
 from lean_spec.forks.lstar.containers import (
     Attestation,
     AttestationData,
+    SignedAttestation,
     SignedBlock,
 )
-from lean_spec.forks.lstar.containers.attestation import SignedAttestation
 from lean_spec.forks.lstar.spec import LstarSpec
 from lean_spec.subspecs.chain.clock import Interval
 from lean_spec.subspecs.chain.config import JUSTIFICATION_LOOKBACK_SLOTS
