+++ tests/lean_spec/forks/lstar/forkchoice/test_compute_block_weights.py
@@ -5,7 +5,7 @@ from __future__ import annotations
 import pytest
 
 from lean_spec.forks.lstar import Store
-from lean_spec.forks.lstar.containers.attestation import AttestationData
+from lean_spec.forks.lstar.containers import AttestationData
 from lean_spec.forks.lstar.spec import LstarSpec
 from lean_spec.subspecs.ssz.hash import hash_tree_root
 from lean_spec.subspecs.xmss.aggregation import TypeOneMultiSignature
