+++ tests/lean_spec/forks/lstar/forkchoice/test_time_management.py
@@ -3,8 +3,7 @@
 from hypothesis import given, settings, strategies as st
 
 from lean_spec.forks.lstar import Store
-from lean_spec.forks.lstar.containers import Block
-from lean_spec.forks.lstar.containers.state import Validators
+from lean_spec.forks.lstar.containers import Block, Validators
 from lean_spec.forks.lstar.spec import LstarSpec
 from lean_spec.subspecs.chain.clock import Interval
 from lean_spec.subspecs.chain.config import (
