+++ tests/lean_spec/helpers/mocks.py
@@ -12,9 +12,11 @@ from dataclasses import dataclass, field
 from types import MappingProxyType
 from typing import Any
 
-from lean_spec.forks.lstar.containers import SignedBlock
-from lean_spec.forks.lstar.containers.attestation import SignedAttestation
-from lean_spec.forks.lstar.containers.attestation.attestation import SignedAggregatedAttestation
+from lean_spec.forks.lstar.containers import (
+    SignedAggregatedAttestation,
+    SignedAttestation,
+    SignedBlock,
+)
 from lean_spec.forks.lstar.spec import LstarSpec
 from lean_spec.subspecs.networking import PeerId
 from lean_spec.subspecs.networking.service.events import NetworkEvent
