+++ tests/lean_spec/subspecs/containers/test_attestation_aggregation.py
@@ -1,6 +1,6 @@
 """Tests for AggregatedAttestation structure."""
 
-from lean_spec.forks.lstar.containers.attestation import (
+from lean_spec.forks.lstar.containers import (
     AggregatedAttestation,
     AttestationData,
 )
