+++ tests/lean_spec/subspecs/networking/client/test_event_source.py
@@ -7,8 +7,7 @@ from unittest.mock import AsyncMock, MagicMock, patch
 
 import pytest
 
-from lean_spec.forks.lstar.containers import SignedBlock
-from lean_spec.forks.lstar.containers.attestation import SignedAttestation
+from lean_spec.forks.lstar.containers import SignedAttestation, SignedBlock
 from lean_spec.snappy import compress
 from lean_spec.subspecs.networking.client.event_source import (
     SUPPORTED_PROTOCOLS,
