+++ tests/lean_spec/subspecs/networking/client/test_gossip_reception.py
@@ -15,8 +15,7 @@ from __future__ import annotations
 
 import pytest
 
-from lean_spec.forks.lstar.containers import SignedBlock
-from lean_spec.forks.lstar.containers.attestation import SignedAttestation
+from lean_spec.forks.lstar.containers import SignedAttestation, SignedBlock
 from lean_spec.snappy import compress, decompress
 from lean_spec.subspecs.networking.client.event_source import (
     GossipHandler,
