+++ tests/lean_spec/subspecs/networking/service/test_service.py
@@ -6,8 +6,7 @@ from __future__ import annotations
 
 from unittest.mock import AsyncMock, MagicMock, patch
 
-from lean_spec.forks.lstar.containers import AttestationData
-from lean_spec.forks.lstar.containers.attestation import SignedAttestation
+from lean_spec.forks.lstar.containers import AttestationData, SignedAttestation
 from lean_spec.snappy import compress
 from lean_spec.subspecs.networking import PeerId
 from lean_spec.subspecs.networking.gossipsub.topic import GossipTopic
