+++ tests/lean_spec/subspecs/sync/test_checkpoint_sync.py
@@ -8,7 +8,7 @@ import httpx
 import pytest
 
 from lean_spec.forks.lstar import State, Store
-from lean_spec.forks.lstar.containers.state import Validators
+from lean_spec.forks.lstar.containers import Validators
 from lean_spec.subspecs.api import ApiServer, ApiServerConfig
 from lean_spec.subspecs.chain.config import VALIDATOR_REGISTRY_LIMIT
 from lean_spec.subspecs.sync.checkpoint_sync import (
