+++ tests/lean_spec/subspecs/sync/test_reaggregate_from_block.py
@@ -24,7 +24,7 @@ from __future__ import annotations
 
 from consensus_testing.keys import XmssKeyManager
 
-from lean_spec.forks.lstar.containers.attestation import AttestationData
+from lean_spec.forks.lstar.containers import AttestationData
 from lean_spec.forks.lstar.spec import LstarSpec
 from lean_spec.subspecs.networking import PeerId
 from lean_spec.subspecs.ssz.hash import hash_tree_root
